blpConnect(RBloomberg)	R Documentation

Open or shut a connection to Bloomberg

Description

The function blpConnect returns a connection object to one of the Bloomberg interfaces. Currently, only the COM Desktop API is supported.

Usage


blpConnect(iface="COM", timeout=12000, show.days="week",
na.action="na", periodicity="daily")
blpDisconnect(conn)

Arguments

iface 	 Which Bloomberg interface? (only "COM" currently supported).

timeout 	 Timeout in Seconds

show.days 	 "trading", "week", or "all"

na.action 	 "bloomberg.handles", "previous.days", or "na"

periodicity 	 "daily", "weekly", "monthly", or "annual"

conn 	 Connection object

Details

Connection to other API's provided by Bloomberg (e.g., C Desktop API and C Server API) will be accessable via this function in future releases of this package. Currently, the object returned is of class 'COMObject'.

## FIX ME ##

Author(s)

Robert Sams rmsams@gmail.com

Examples


## by default establish connection via Desktop COM API
## Not run: 
conn <- blpConnect()
blpDisconnect(conn)
## End(Not run)

[Package RBloomberg version 0.2-2 Index]