blpReadFields(RBloomberg)	R Documentation

Read bbfields.tbl file into workspace.

Description

Read into the workspace the file bbfields.tbl located on the local machine.

Usage


blpReadFields(path = "C:/blp/API")

Arguments

path 	 character. path where bbfields.tbl can be found.

Details

In order to perform datatype mapping, RBloomberg needs to consult the Bloomberg API file bbfields.tbl located on the machine running the Bloomberg terminal. We don't physically search the file each time a datatype mapping is needed. Instead, we read the entire contents of the file into the workspace and store it in a hidden dataframe .bbfields.

This is attempted when the package is first loaded. If the file is not found in C:/blp/API, you must manually specify the correct path using this function.

Author(s)

Robert Sams rmsams@gmail.com

Examples


.bbfields <- blpReadFields()

[Package RBloomberg version 0.2-2 Index]